+++ src/main/scala/io/github/mandar2812/dynaml/examples/TestAdult.scala
@@ -22,12 +22,13 @@ import java.io.File
 
 import breeze.linalg.{DenseMatrix, DenseVector => BDV}
 import com.github.tototoshi.csv.CSVWriter
+import io.github.mandar2812.dynaml.DynaMLPipe
 import io.github.mandar2812.dynaml.evaluation.BinaryClassificationMetrics
 import org.apache.spark.mllib.regression.LabeledPoint
 import org.apache.spark.rdd.RDD
 import org.apache.spark.{SparkConf, SparkContext}
 import io.github.mandar2812.dynaml.kernels.{RBFKernel, SVMKernel}
-import io.github.mandar2812.dynaml.models.KernelizedModel
+import io.github.mandar2812.dynaml.models.{GLMPipe, KernelizedModel}
 import io.github.mandar2812.dynaml.models.lm.GeneralizedLinearModel
 import io.github.mandar2812.dynaml.models.svm.{KernelSparkModel, LSSVMSparkModel}
 import io.github.mandar2812.dynaml.pipes._
