+++ src/main/scala/io/github/mandar2812/dynaml/examples/TestCommitteeNNOmni.scala
@@ -22,9 +22,10 @@ package io.github.mandar2812.dynaml.examples
 
 import breeze.linalg.DenseVector
 import com.quantifind.charts.Highcharts._
+import io.github.mandar2812.dynaml.DynaMLPipe
 import io.github.mandar2812.dynaml.evaluation.RegressionMetrics
 import io.github.mandar2812.dynaml.models.neuralnets.{CommitteeNetwork, FFNeuralGraph}
-import io.github.mandar2812.dynaml.pipes.{DynaMLPipe, DataPipe}
+import io.github.mandar2812.dynaml.pipes.DataPipe
 import org.apache.log4j.Logger
 
 /**
