+++ src/main/scala/io/github/mandar2812/dynaml/examples/TestGPHousing.scala
@@ -19,11 +19,13 @@ under the License.
 package io.github.mandar2812.dynaml.examples
 
 import breeze.linalg.{DenseMatrix, DenseVector}
+import io.github.mandar2812.dynaml.DynaMLPipe
 import io.github.mandar2812.dynaml.evaluation.RegressionMetrics
 import io.github.mandar2812.dynaml.kernels._
+import io.github.mandar2812.dynaml.models.GPRegressionPipe
 import io.github.mandar2812.dynaml.models.gp.GPRegression
 import io.github.mandar2812.dynaml.optimization.{GPMLOptimizer, GridSearch}
-import io.github.mandar2812.dynaml.pipes.{BifurcationPipe, DataPipe, DynaMLPipe, GPRegressionPipe}
+import io.github.mandar2812.dynaml.pipes.{BifurcationPipe, DataPipe}
 
 /**
   * Created by mandar on 15/12/15.
