+++ src/main/scala/io/github/mandar2812/dynaml/examples/TestHousing.scala
@@ -2,7 +2,7 @@ package io.github.mandar2812.dynaml.examples
 
 import com.tinkerpop.blueprints.Graph
 import com.tinkerpop.frames.FramedGraph
-import io.github.mandar2812.dynaml.graphutils.CausalEdge
+import io.github.mandar2812.dynaml.graph.CausalEdge
 import io.github.mandar2812.dynaml.models.KernelizedModel
 import io.github.mandar2812.dynaml.models.svm.LSSVMModel
