+++ src/main/scala/io/github/mandar2812/dynaml/examples/TestRipley.scala
@@ -5,7 +5,7 @@ import java.io.File
 import com.github.tototoshi.csv.CSVWriter
 import com.tinkerpop.blueprints.Graph
 import com.tinkerpop.frames.FramedGraph
-import io.github.mandar2812.dynaml.graphutils.CausalEdge
+import io.github.mandar2812.dynaml.graph.CausalEdge
 import io.github.mandar2812.dynaml.models.KernelizedModel
 import io.github.mandar2812.dynaml.models.svm.LSSVMModel
