+++ src/main/scala/io/github/mandar2812/dynaml/graph/ParamEdge.scala
@@ -16,22 +16,19 @@ KIND, either express or implied.  See the License for the
 specific language governing permissions and limitations
 under the License.
 * */
-package io.github.mandar2812.dynaml.graphutils
+package io.github.mandar2812.dynaml.graph
 
-import com.tinkerpop.blueprints.Direction
-import com.tinkerpop.frames.{Adjacency, VertexFrame, Property}
+import com.tinkerpop.frames.{InVertex, OutVertex, EdgeFrame}
 
 /**
- * Defines the [[VertexFrame]] for the
- * parameter node in the graph.
+ * Defines the [[EdgeFrame]] for the
+ * edges going out from the parameter
+ * node.
  */
-trait Parameter extends VertexFrame {
-  @Property("slope")
-  def getSlope(): Array[Double]
+trait ParamEdge extends EdgeFrame {
+  @OutVertex
+  def getParameter(): Parameter
 
-  @Property("slope")
-  def setSlope(slope: Array[Double]): Unit
-
-  @Adjacency(label = "controls", direction = Direction.OUT)
-  def getControlledPointLabels(): java.lang.Iterable[Label]
+  @InVertex
+  def getLabel(): Label
 }
