+++ src/main/scala/io/github/mandar2812/dynaml/graph/Point.scala
@@ -16,22 +16,28 @@ KIND, either express or implied.  See the License for the
 specific language governing permissions and limitations
 under the License.
 * */
-package io.github.mandar2812.dynaml.graphutils
+package io.github.mandar2812.dynaml.graph
 
 import com.tinkerpop.blueprints.Direction
 import com.tinkerpop.frames.{Adjacency, VertexFrame, Property}
 
 /**
  * Defines the [[VertexFrame]] for the
- * target label node in the graph.
+ * input features nodes in the graph.
  */
-trait Label extends VertexFrame{
+trait Point extends VertexFrame {
   @Property("value")
-  def getValue(): Double
+  def getValue(): Array[Double]
 
   @Property("value")
-  def setValue(value: Double): Unit
+  def setValue(value: Array[Double]): Unit
 
-  @Adjacency(label = "causes", direction = Direction.IN)
-  def getCausedByPoint(): java.lang.Iterable[Point]
+  @Property("featureMap")
+  def getFeatureMap(): Array[Double]
+
+  @Property("featureMap")
+  def setFeatureMap(value: Array[Double]): Unit
+
+  @Adjacency(label = "causes", direction = Direction.OUT)
+  def getLabel(): java.lang.Iterable[Label]
 }
