+++ src/main/scala/io/github/mandar2812/dynaml/graph/Synapse.scala
@@ -16,25 +16,37 @@ KIND, either express or implied.  See the License for the
 specific language governing permissions and limitations
 under the License.
 * */
-package io.github.mandar2812.dynaml.graphutils
+package io.github.mandar2812.dynaml.graph
 
-import com.tinkerpop.frames.{Property, OutVertex, InVertex, EdgeFrame}
+import com.tinkerpop.frames.{Property, InVertex, OutVertex, EdgeFrame}
 
 /**
- * Defines the [[EdgeFrame]] for the
- * edges going out from the input data
- * node.
+ * Frames implementation of a Neural Network
+ * Synapse
  */
-trait CausalEdge extends EdgeFrame {
+trait Synapse extends EdgeFrame {
   @OutVertex
-  def getPoint(): Point
+  def getPreSynapticNeuron(): Neuron
 
   @InVertex
-  def getLabel(): Label
+  def getPostSynapticNeuron(): Neuron
 
-  @Property("relation")
-  def getRelation(): String
+  @Property("weight")
+  def getWeight(): Double
 
-  @Property("relation")
-  def setRelation(value: String): Unit
+  @Property("weight")
+  def setWeight(w: Double): Unit
+
+  @Property("PrevWeightUpdate")
+  def getPrevWeightUpdate(): Double
+
+  @Property("PrevWeightUpdate")
+  def setPrevWeightUpdate(w: Double): Unit
+
+  @Property("layer")
+  def getLayer(): Int
+
+  @Property("layer")
+  def setLayer(l: Int): Unit
+
 }
