+++ src/main/scala/io/github/mandar2812/dynaml/models/ModelPipe.scala
@@ -1,0 +1,14 @@
+package io.github.mandar2812.dynaml.models
+
+import io.github.mandar2812.dynaml.pipes.DataPipe
+
+/**
+  * Top level trait for Pipes involving ML models.
+  */
+trait ModelPipe[Source, T, Q, R, M <: Model[T, Q, R]]
+  extends DataPipe[Source, M]{
+
+  val preProcess: (Source) => T
+
+  override def run(data: Source): M
+}
