+++ src/main/scala/io/github/mandar2812/dynaml/models/ensemble/CommitteeModel.scala
@@ -21,8 +21,7 @@ package io.github.mandar2812.dynaml.models.ensemble
 import breeze.linalg.DenseVector
 import io.github.mandar2812.dynaml.models.gp.GPRegression
 import io.github.mandar2812.dynaml.models.neuralnets.FeedForwardNetwork
-import io.github.mandar2812.dynaml.models.{LinearModel, Model}
-import io.github.mandar2812.dynaml.pipes.ModelPipe
+import io.github.mandar2812.dynaml.models.{LinearModel, Model, ModelPipe}
 
 /**
   * Defines an abstract implementation of a "committee-model".
