+++ src/main/scala/io/github/mandar2812/dynaml/models/neuralnets/FFNeuralGraph.scala
@@ -21,7 +21,7 @@ package io.github.mandar2812.dynaml.models.neuralnets
 import breeze.linalg.DenseVector
 import com.tinkerpop.blueprints.{GraphFactory, Graph}
 import com.tinkerpop.frames.{FramedGraphFactory, FramedGraph}
-import io.github.mandar2812.dynaml.graphutils.{Neuron, Synapse}
+import io.github.mandar2812.dynaml.graph.{Neuron, Synapse}
 import org.apache.log4j.Logger
 
 import scala.collection.JavaConversions
