+++ src/main/scala/io/github/mandar2812/dynaml/models/svm/LSSVMModel.scala
@@ -8,7 +8,7 @@ import com.tinkerpop.blueprints.util.io.graphson.GraphSONWriter
 import com.tinkerpop.blueprints.{Graph, GraphFactory}
 import com.tinkerpop.frames.{FramedGraph, FramedGraphFactory}
 import io.github.mandar2812.dynaml.evaluation.Metrics
-import io.github.mandar2812.dynaml.graphutils.{CausalEdge, Label, ParamEdge, Parameter, _}
+import io.github.mandar2812.dynaml.graph.{CausalEdge, Label, ParamEdge, Parameter, _}
 import io.github.mandar2812.dynaml.optimization._
 import io.github.mandar2812.dynaml.utils
 import org.apache.log4j.{Logger, Priority}
