+++ src/main/scala/io/github/mandar2812/dynaml/models/svm/SDLSSVMModel.scala
@@ -4,7 +4,7 @@ import breeze.linalg.{DenseVector, DenseMatrix}
 import com.tinkerpop.blueprints.{GraphFactory, Graph}
 import com.tinkerpop.frames.{FramedGraphFactory, FramedGraph}
 import io.github.mandar2812.dynaml.evaluation.Metrics
-import io.github.mandar2812.dynaml.graphutils._
+import io.github.mandar2812.dynaml.graph._
 import io.github.mandar2812.dynaml.kernels.SVMKernel
 import io.github.mandar2812.dynaml.models.SubsampledDualLSSVM
 import io.github.mandar2812.dynaml.optimization.ConjugateGradient
