+++ src/main/scala/io/github/mandar2812/dynaml/optimization/BackPropogation.scala
@@ -19,7 +19,7 @@ under the License.
 package io.github.mandar2812.dynaml.optimization
 
 import breeze.linalg.DenseVector
-import io.github.mandar2812.dynaml.graphutils.Neuron
+import io.github.mandar2812.dynaml.graph.Neuron
 import io.github.mandar2812.dynaml.models.neuralnets.FFNeuralGraph
 import org.apache.log4j.Logger
