+++ src/main/scala/io/github/mandar2812/dynaml/optimization/GradientDescent.scala
@@ -19,7 +19,7 @@ under the License.
 package io.github.mandar2812.dynaml.optimization
 
 import breeze.linalg.DenseVector
-import io.github.mandar2812.dynaml.graphutils.CausalEdge
+import io.github.mandar2812.dynaml.graph.CausalEdge
 import io.github.mandar2812.dynaml.pipes.DataPipe
 import org.apache.log4j.{Logger, Priority}
